(SERVER-1823) Remove certname from http client metrics
When metrics gathering was added for PuppetDB commands, the certname
associated with the command was included as part of the metrics path.
This can cause a large number of individual metrics to be generated in
large environments, which is undesirable for a few reasons. By itself
the number of metrics can be unwieldy, and there's memory overhead for
storing each of these metrics.

This commit reduces the metric proliferation by dropping the certname
from the metrics path when an HTTP request is made.