(PDB-3548) travis: use jobs, not matrix; don't use language
Stop specifying "language: clojure" so that we can control if/when
things like "lein deps" happen.  Aside from making the process
clearer, this also means that we can make sure we're only doing the
work necessary for each run, i.e. we don't need, and don't want to run
lein deps, or to install puppetserver for the ruby spec tests.

Select "language: c" because it's likely to be fairly minimal, and
orthogonal enough to reduce potential conflicts.  The resulting
environment still has the clojure and ruby bits needed.

Switch from matrix includes to jobs so that we can specify exactly
what each test should do, and move more of the things that aren't
travis specific out of the travis-specific file.