(PDB-3549) Test: set pkg ver to 'latest' if undef
This commit updates the acceptance helper method `get_package_version`
to return the 'latest' string if the puppetdb version was not defined
by another means previously.

If the package version is set to 'latest' then the validation for
matching the version is skipped since it does not make sense in this
case.

The pre-suite step for installing the puppetdb dev repo is skipped if
the build version is not specified.