(SERVER-1823) Remove Http.multi_arity_get
Per PE-18347 earlier versions of Puppet and Puppetserver had different
HTTP clients whose arity of `#get` was not consistent.
`Http.multi_arity_get` was introduced to resolve this issue of varying
functionarities and would select the appropriate http client invocation
based on the arity of the underlying method. In Puppetserver 5 the HTTP
client implementation was changed to match that of Puppet rendering the
`.multi_arity_get` method unneeded; this commit removes the use of that
method in favor of directly calling `#get` on the http client instance.