(PDB-3525) Trigger extensions in final travis build stage
Use travis build stages to only trigger the extensions build if all of
the tests succeed.  Otherwise I think we'd trigger the extensions
build once per matrix item.

The previous matrix did not appear to work well with build stages, so
rewrite the matrix entries (which were already explicitly listed) to
an explicit job list, and move more of the things that aren't travis
specific out of the travis-specific file.