(PDB-3195) Default transaction isolation in connection pool config
Before this change, our internal transaction wrapper macro was explicitly
passing a :read-committed isolation level to clojure.jdbc when not otherwise
specified. This causes a call to .setTransactionIsolation on the jdbc driver,
which executes SQL on the server. So it ends up making an extra network round
trip for each query.

This commit sets read-committed transaction isolation in the connection pool
config, so each connection already has that by default. It also allows nil
isolation levels to be passed to clojure.jdbc, which will then use whatever the
connection already has.