(RE-8458) Update tests to use the puppet repos
In order to test against the right set of packages, we need to update
these tests to hit the new repos. Rather than pin to puppet 5, we're
just going to pin to the rolling puppet repos. This will change
automatically to puppet6 once we change these repos to point to puppet6.