(PDB-3648) Rewrite null json chars via jackson for pg
Instead of rewriting the final json string to replace null (\u0000)
chars with the replacement char (\ufffd), specify a custom
CharacterEscapes class to the jackson json factory that will handle
the conversion of any null characters directly, during the encoding.

The previous approach might not always produce the correct output.
For example, it could end up generating a json string containing a
backslash followed by the replacement character, which is invalid,
and which postgres will reject.