(PDB-3686) Handle duplicate null rows in facts migration
We have observed that some databases can end up with two kinds of 'null' values:
real sql null, and json null. (that is, a null value encoded as json in the
value column and stored with the json type) Since these are hashed the same,
this causes an error in the facts migration.

This change finds and duplicates and keeps the real (sql) null one, rewriting
any facts that used the json null.