(PDB-3734) Ensure that resource indirection uses existing resource type
Before this commit, the resource type would be passed to `Resource#new`
in string form. This is problematic since it may result in the load of
a Ruby resource type which then is in conflict with genereated Pcore
resource type.

This commit ensures that the indirection uses the same logic as Puppet
when creating new resources.