(PDB-3745) Reduce autovacuum vacuum scale factor on reports
Prior to this commit, the reports table was not vacuumed
frequently enough under default postgresql settings.

After this commit, we start looking to vacuum after 1% of the table
changes.  For 1000 nodes, checking in every 30 minutes and storing
14 days of reports we'd have ~ 336,000 reports.  Every 30 minutes
1000 new reports show up.  With a 1% autovacuum scale factor we
look to vacuum after every 3,360 reports which is essentially every
2 hours ( assuming a 1 hour gc-interval).  In the case that you
store less than the default 14 days of reports then the reports
table will be vacuumed more frequently, however, we know it will
only happen after waiting gc-interval since that's how often reports
are deleted.

Long story short, assuming a default gc-interval of 1 hour we know
that we can't force autovacuum on reports any more frequently than
once per hour and with default report-ttl and runinerval we'd expect
report GC every 2 hours.  This is good because the reports table
is the largest table and we want to keep the size down by frequently
vacuuming it.