(PDB-3723) Add garbage collection for jsonb derived fact_paths
Remove fact paths that aren't referenced by any jsonb factset during
the normal garbage collection process.  Note that the candidates are
selected solely based on the path, ignoring the type.  As a result, if
a path changes type, the path with the older type will remain
alongside the new one.

Thanks to Wyatt Alt for providing the foundation of the recursive
live_paths query.