(PDB-3311) Improve error handling for index row size exception
Sometimes users make a puppet error where they try to extract a fact
value with $facts["foo"] instead of ${facts["foo"]}, which causes the
entire factset to be embedded in the catalog. If the factset is large,
this results in data that is too big for a postgres btree index, which
produces a confusing error message.

Intercept those error messages and print something more comprehensible
and divert those messages straight to the DLO, since retrying them
will always fail.