(PDB-3838) Optimize fact queries which constrain to a single name
Previously, all queries to the facts endpoint generated brute-force
fallback sql which extracted all fact names and values using
jsonb_each and filtered the resulting tuple. This is very slow.

This change optimizes queries which constrain the fact name to a
single value, the most common way of using the facts endpoint. It does
this by rewriting the facts query-rec to explicitly extract the
indicated value using the jsonb -> operator.

[rlb@puppet.com: rewrap commit message and a docstring; change
 constrains-name? to name-constraint; reword comment describing
 :selection-params]