(PDB-3852) Fix broken structured facts queries
This commit fixes structured facts queries that descend into arrays and
those that contain regexes.

Regexes on structured facts queries have been fixed by querying the
database for all paths that match the given path regex and then adding
the results from that into the original user-query with an "or"
statement around the returned paths.

Queries that descend into arrays have been fixed by examining the query
and using the #> operator instead of the @> operator in the where
clause.