(PDB-3911) Add primary key to resource_events
Prior to this commit, resource_events had a unique constraint
on four columns.

After this commit, we add a primary key on those same four columns
and drop the unique constraint.  A primary key is enforced by a
unique constraint behind the scenes so there should be no difference
other than 3rd party tools like pglogical or pg_repack will now
be happy that there is a primary key when they used to fail.