(PDB-3917) Stop mangling project.clj via sed for integration tests
Move the install-gems code to src-gems and manage it via an
install-gems profile so that the puppetserver dependency doesn't
always have to be known (install_gems.clj refers to puppetserver
namespaces and so can't be compiled without a puppetserver
dependency).

Making the puppetserver dependency optional is important because until
we actually check out the ./puppetserver tree, we don't know what the
version is, i.e. for 5.3.x it might be 5.3.7-SNAPSHOT.

Have configure-int-tests store the puppetserver test dependency
version in ext/test-conf/puppetserver-dep, and adjust everything to
read the version from there when the file exists, or behave reasonably
when it doesn't.

Previously, not knowing the real depdendency version hasn't been an
issue because we've set puppetserver-version in the project.clj to
some fixed version, and then mangled the project.clj via sed to adjust
it during integration testing.