(PDB-3963) Use proper gem-list from puppetserver
Prior to this commit (due to a file name change in puppetserver), the
integration tests were broken between master branches.

This commit makes it possible for the puppetdb master branch to run
against master as well as the tagged releases of puppetserver by
choosing the proper file name depending on puppetserver's version.