(PDB-3268) Don't migrate on schema_migration query failures
This commit, if the schema_migrations query fails, will only attempt the
migrations if it fails due to an `undefined_table` error. In this case,
the database has not been initialized and so all migrations will be
applied. In all other cases, the error will be thrown.

This was the original intention of the error-handling here as per
commit fc0252c (PR #1707)