(PDB-3948) Fix transient failure of pretty print
Previously, it was possible for garbage collection to run in the middle
of the developer-pretty-print test. which if it ran in between adding
the cert and facts and the checks would clean them out of the database.

This commit changes the test to use the timestamp now so that even if
garbage collection runs, the facts will remain.