(PDB-3470) Add pdbbox and use it in tests; test osx in travis
See the documentation/ and .travis.yml changes for further details.

Note that this adds additional inter-dependent utility scripts to
ext/bin, but we don't want to add ext/bin to the PATH because so far,
the command names in ext/bin haven't been chosen with that in
mind (i.e. sha256sum would shadow the Linux version), so we refer to
them with explicit paths (sometimes generated via basename).

In the longer run we might want to adjust the relevant command names,
or perhaps move "utility" commands to ext/util/, to make ext/bin
suitable for inclusion in the PATH.

Remove the ulimit -u settings because it removes another platform
variance (disallowed on travis osx hosts), and it looks like we may no
longer need it, but if we see the OutOfMemoryError exceptions
return (2080a332b40f95cdf564a65a6bf540704273b583), we can restore it.