(PDB-3993) Only su in foreground subcommand if id isn't already USER
So if you're already the right user, say because you already were, you
wanted to be explicit about the use of su, or you wanted to use
something else like sudo, then you can.  i.e. so

  su -l puppetdb puppetdb foreground

will avoid trying to su again.

There are other ways we might handle this if we were starting from
scratch, but this approach attempts to minimize the changes for now.