(PDB-4020) Return the last successful reponse with command_broadcast
Prior to this commit, the http submission with command_broadcast enabled
always returned the last response. As a result, a failure would be shown if
the last connection produced a 503 response even though there was
previously a successful PuppetDB response and the minimum successful
responses have been met. This issue does not occur with responses that
raised an exception. Since the puppet http_pool does not raise 503
as an exception, this issue can be seen when the PuppetDB is in
maintenance mode.

This commit changes the behavior to send the last successful response
when the minimum successful submissions have been met.