(PDB-3993) Add pdb upgrade subcommand (upgrades and exits)
This should at least be useful for cases where you need to upgrade
across multiple major versions without skipping major versions (as per
versioning policy) e.g:

  ...pdb-5/bin/puppetdb upgrade || die  # Migrate from 4 to 5
  ...pdb-6/bin/puppetdb upgrade || die  # Migrate from 5 to 6
  ...

To coordinate across multiple services, insinuate a [global]
upgrade-and-exit? option into the config, and check it from the
relevant services.  Leave puppetdb in maintenance mode during the
process, and have the the command service turn off the
lights (i.e. call trapperkeeper request-shutdown) once it's finished
the queue migration, since we know everything else has been done at
that point.

The way this is handled isn't all that pretty, but it's a relatively
simple approach with a small footprint, and should be removed (along
with activemq itself in version 6).