(PDB-4108) Fix AOT issues and compile uberjar
Thanks to Deepak Giridharagopal for help working around the
unexpected OptionalKey behavior, which was causing exceptions like
this during compilation:

  Caused by: clojure.lang.ExceptionInfo: Input to kwd->str does not match schema:
             [(named (not (keyword? a-schema.core.OptionalKey)) kwd)]

We observed that and (OptionalKey. ...) instance created at the point
of failure did not have the same (class x), hashCode-wise, as the
instance coming in to the function performing the instance? check.

Comparing the classes with = appears to side-step the problem.

As part of this work, drop the "testutils" jar -- it's not clear
anyone was using it, it was causing trouble with AOT, and we can
always pursue those issues and restore it, if it turns out theres a
demand.