(PDB-2001) Raise an error when an invalid date is sent on a command
This adds validation of the producer_timestamp value inside a command body.

This supports all the common date-time formats, and adds support for UNIX
timestamps as strings (as is used in a lot of tests)

When validation fails, the server returns an HTTP 500, matching the other
behavior of failures for missing required parameters and such.