(PDB-2001) Validate producer-timestamp in query string
Adds validation to the producer-timestamp in the query string of a command. This
ignores the producer_timestamp in the body, as we do not do validation of
command bodies until the command is popped off the queue.

Also modifies tests that submit producer_timestamp to reflect the proper 'Z'
format of ISO-8601. The command processors fail when this value is a UNIX
timestamp instead of an ISO-8601 string, so updating the tests improves our
"example" code for this field to match the wire format.