(PDB-4165) Duplicate dependency pins in :ezbake profile
It looks like the use of ^:replace :dependencies in the ezbake profile
causes all pins in the normal :dependencies section to be ignored.  So
for now, just repeat the pins for the recent CVEs in the :ezbake
profile, which does cause them to show up here:

  lein with-profiles ezbake deps :tree

In addition, stop pinning hikaricp and clojure/jdbc so that we won't
fall afoul of :pedantic, *and* so that we'll actually be testing the
versions that we ship (verified by spot-checking the relevant
META-INF/maven/ pom.xml versions in the 5.1.5 debian/stretch jar).