(PDB-3742) Add ability to track enqueued commands
This commit adds the ability to track commands which are enqueued but
not yet processed. This is accomplished by sending maps which represent
the state of ingested commands on a channel which is consumed and used
to update a tracking atom on the pdbext side. By keeping track of what
commands are currently enqueued we are able to avoid extra work during
pdbext HA sync.