(maint) Update puppetdb entrypoint for less dependence on puppetserver
Previously, on anything but the initial run, if the puppetdb container
started up before puppetserver, the container would fail and exit since
the directory that is checked for existence never exists on startup. By
updating the checks in the entrypoint, there is now no hard dependency
between puppetserver and puppetdb.