(MAINT) Fix cacert for adoptopenjdk10
The previous commit fixed the ability to download adoptopenjdk10 but the
cacert was still not compatible with clojars, so when a dependency
changed, it could not be downloaded.

This commit fixed adoptopen10's cacert in the same way we fixed it for
adoptopenjdk8, by downloading 11 and using its fixed cacert.