(PDB-4199) Allow multi-value array in 'in' queries
Previously, when querying base on fact values using the 'in' operator,
the query would fail for any array other than a singleton array.
This fixes the type checking by using a set to de-depulicate the array
of fact-value types.

This also updates a documentation example that is no longer valid due to
a stricter type-checking.