(maint) Allow configuration of certname in ssl-setup.sh
Previously the `ssl-setup.sh` script would use the `HOSTNAME` envioronment when dealing with `{certname}.*.pem` ssl cert paths. In the case where a hard coded cert name is copied in to place in a Dockerfile that derives from the puppetdb image allow the configuration of the certname with an environment variable called `CERTNAME`. The default of `HOSTNAME` will still be respected in the case that `CERTNAME` is unset.