(MAINT) Add container analytics (#2804)
This was originally b639fb6c95f95533620bd60355c250517d569a5b and should
have been included in the mergeup from 5.2.x -> 6.0.x as part of commit
3724af623066a409be896c5471867244c1ee697a. However, it was mistakenly
not included and has been cherry-picked forward.

Add an entrypoint script to submit a "start" event to our pupperware
account on Google Analytics.

The script will generate a UUID and reuse it as an anonymous client ID,
in an attempt to more accurately identify the number of unique instances
of the container. This UUID file is kept in /var/tmp, which doesn't
require explicit Docker volumes and is persisted across container
restarts.

The `ec` field is for the Event Category, which can either be "dev" or
"production", and is a configured as a container build argument. Only
the containers that we ship to DockerHub should be built as
"production", and the default is "dev".

Tests should not report as either though, so we specify
PUPPERWARE_DISABLE_ANALYTICS when running tests.

Users can disable analytics using a non-empty value for
PUPPERWARE_DISABLE_ANALYTICS, and also specify which stream to report as
using PUPPERWARE_ANALYTICS_STREAM=sometext.