(PDB-4257) Don't mangle already mangled queue metadata
Add an fs-ready? argument to the metadata-serializer that indicates
whether or not the data has already been prepared for the
filesystem (i.e. has already been sanitized and/or truncated) so that
it won't mangle already mangled data to produce metadata that won't
match the actual queue filenames.

Previously, when trying to read an existing queue file, pdb would just
blindly mangle the the data in the command ref (if needed due to
length or the existence of forbidden characters), even though that
data should already have been mangled.

As a result it could end up asking stockpile for an incorrect entry,
and the visible effect was that pdb would claim in the logs that the
command had "disappeared".

As an example of the problem, pdb would (and should, because we
reserve underscore as our own delimiter in queue metadata) mangle a
certname of foo_bar like this:

  foo-bar_5d5f20e74771b852025c3edb66c9462eeb913d03

which would become the command ref's :certname (obviously not the real
certname) and a component of the command's queue filename.

But then, when preparing to process that command, pdb didn't know that
the "certname" had already been mangled, and so would end up seeing
the "_" and mangle it again, producing this:

  foo-bar_5d5f20e74771b852025c3edb66c9462eeb913d03_29baf23de6424a39e4dc60b46fce28b9f2c06e53

...which of course doesn't actually match the original file in the
queue.

Temporarily change the certname in the concurrent-fact-updates test to
avoid mangling so that all the tests will still pass -- until we fix
the (remaining) issue with parsing metadata containing a mangled
certname.