(PDB-4277) Add logback logstash encoder
This commit adds a dependency on net.logstash.logback/logstash-logback-encoder
which is described in the PuppetDB documentation as a tool for transforming
log messages into JSON format:

  https://puppet.com/docs/puppetdb/6.2/logging.html#json-text

Prior to this commit, the required library was missing from the uberjar. The
library is managed by clj-parent as Puppet Server includes it.