(PDB-4271) query for fact expiration
Adds an "expires" field to the node response, that contains two embedded
fields: "facts" and "facts_updated"

If "facts" is true (the default), the node is subject to garbage
collection and expiration. If the value is set to false, the node is
excluded from the garbage collection procedure. If this value has been
explicitly set via a command, the facts_updated field will be a
timestamp representing when the field was set.