(SERVER-2497) Extract `job_id` from request metadata
Previously, the `save` method of the catalog terminus would only save
`job_id` if it came in as part of the catalog. This commit ensures that
we also extract `job_id` from the extra request data if it is present,
and use that if `job_id` is not present in the catalog.