(SERVER-2497) Convert job_id to string
PuppetDB has a schema in place for the wire format of the catalog that
requires `job_id` to be a string. Since it comes from Puppet as an
integer, this commit converts it to a string in the terminus before
passing it on to PDB proper. If no `job_id` was supplied with the
request, it is left nil.