(PDB-4312) handle duplicates in resource_events table
The resource_events table can contain duplicate rows because the existing
unique constraint does not work as intended. In PostgreSQL, null values
are not considered equal in an index.

Migration now uses a set to dedupe the resource_events table as its being
written to, and the unique constraint is dropped in favor of the new
primary key on the resource_events table.