(PDB-4318) Fix pre_suite steps getting skipped
Prior to this commit the logic for deciding whether to run pre_suite steps 70
and 80 was wrong and resulted in :upgrade tests not running all the intended
steps. This commit also changes the order of the 70 and 80 pre_suite steps to
avoid adding the dev repos until after the released PuppetDB has been installed.