(PDB-4326) Allow unicode in resource tags
This changes the tag regex in PuppetDB to allow all non-uppercase unicode
letters as tags. It will also reject tags ending in newlines.

In PUP-7579, Puppet added support for Unicode letters in resource tags,
but the supported tags regex was never changed in PuppetDB so any
catalogs using unicode resource tags would not be stored in PuppetDB.

The current regex is intended to match the ruby patten
`\A[[:alnum:]_][[:alnum:]_:.-]*\Z`, which takes advantage of the fact
that Ruby is using Unicode character classes to determine what is
alphanumeric.