(PDB-4308) Don't repeatedly open/close db pool for migrations
Rework init-with-db to only open one connection pool, outside the
loop, since we shouldn't need more than one, and simplify the code
somewhat.  Change the return value to unspecified, since it wasn't
used anywhere and doing so simplifies the code.

Remove the bit in the make-connection-pool docstring about it timing
out.  Calls to that function don't time out in 30000ms (HikariCP
default).  It's any attempt to get a connection from the pool it
returns that do.