(maint) Update Makefile / .travis.yml to use rvm
Bundle installing / bundle execing with an RVM project in travis gets
complicated when there's a Gemfile at the root of your repo and you're
trying to use a gemfile from a different directory.

This makes sure we're setting the BUNDLE_PATH and BUNDLE_BIN environment
variables, and explicitly passes the path for the Gemfile for all bundle
commands.