(PDB-4390) Store the uncoerced catalog certname
If a certname contain invalid characters for the queue metadata, such as
an underscore, or if it is too long it will be coerced and a hash of the
original will be appended to the end. For replace-catalogs, this
certname would end up getting stored in the database (only after PDB-4257).

Now the certname is correctly read out of the command payload (the
uncoerced value) and stored properly.