travis: use adoptopenjdk for linux to avoid travis churn
Since travis keeps changing the jdk versions it supports and/or how
you're allowed to access them, switch to adoptopenjdk for all versions
on linux so we'll hopefully be interrupted less often.  This should
fix recent build failures on some branches.

Now that we don't need jdk_switcher (which was a shell function so we
couldn't invoke it from sub-commands), move all the os-specific work
to prep-os-essentials-for, and allow for it to create a shell
snippet that can be sourced by the caller to handle any necessary
environment changes.

Switch to xenial from trusty so that all the desired major versions of
the jdk will be available.  Drop oraclejdk since it isn't available
for xenial.