(PDB-4266) Add include_package_inventory flag to factsets query
This adds support for the include_package_inventory query option, which
when provided will re-create the original wire format of
package_inventory. This is intended to support proper syncing of
factsets in PE.

For what we tested as "average" nodes (ones with roughly 2-5 thousand
packages) we noticed a query slow-down of about 5x when the new
include_package_inventory flag is provided (but the query returns about
21x the data so this is somewhat expected). This may increase the load
of PuppetDB during sync, but will actually sync the required data to get
to a consistent state.

The changes to the tests required changing the expected hashes of
factsets because the package_inventory was previously omitted from the
submitted data, so now a different set of data is being hashed for the
test factsets.