(PDB-4420) partition of resource_events
Prototype of partitioning the resource_events table. This creates a base
table (from the same schema as the existing resource_events table) and
will create 8 total partitions during migration (this ISO week +- 4)

Partitions are based on ISO 8601 weeks. Inserting into the partitions is
controlled via a trigger that directs the insert to the appropriate
partition.

Partitions are dynamically created by scanning the events to be inserted
and performing a "CREATE TABLE IF NOT EXISTS" for the partition. This is
akin to the (ensure-certname) code that already exists.

The inheritance-based implementation was chosen to maintain backwards
compatibility with PostgreSQL 9.x, since the other components of PE are
not ready for PostgreSQL 11.

TODO: Migrate existing events to partitioned tables - they are currently
dropped.

TODO: GC by dropping expired partitions.