(PDB-4266) Remove package_inventory key from hash if empty
If the package inventory is empty, we don't store anything in the
database, and it is therefore undifferentiable from a factset that
didn't have a package_inventory submitted with it at all. This would
create a problem when syncing by potentially generating the wrong hash
when we couldn't tell if we were syncing an empty package inventory or a
missing package inventory.

By removing an empty package_inventory key, we generate a hash that is
consistent and reproducible from the data stored in PostgreSQL.