(PDB-4372) Accept replace catalog-inputs command
Add the "replace catalog-inputs" command to accept storing the names of
catalog inputs, like hiera keys. Although they are associated with
catalogs, they are stored as a component of a node because they may not
be calculated by every catalog compilation and they need to remain
accessible if that happens.

Catalog inputs are garbage collected when a node is garbage collected
via an ON DELETE CASCADE foreign key.