(PDB-4373) redhat: kill old pdb in preinst given 6.3 main ns change
pdb's main namespace was changed in 6.3, during the work to compile
the uberjar, and we didn't realize that ezbake captures and depends on
it[1].  (It changed from "puppetlabs.puppetdb.main" to
"puppetlabs.puppetdb.cli.services".)

That alone might not be an issue, but an acceptance test failure
suggests that our rpms are using the new version's namespace[1] when
trying to kill the old server during an upgrade (perhaps via the new
version's cli/stop script).

From looking at the rpm --scripts for the current 5.2 rpm, it appears
that we don't stop and then later start pdb during an upgrade, we only
restart it, and we only do that once, after the new package is
installed, from the old package's "%postun" (postuninstall) script --
step 11 here:

  https://docs.fedoraproject.org/en-US/packaging-guidelines/Scriptlets/#ordering

At that point, assuming the cli/stop scripts, etc. are those from the
new package, then the pgrep will be looking for the wrong namespace to
kill and the old server won't be stopped.

To fix that, hack the new pdb rpm postinst (presumably run at step 4
here) to always try to kill the old server (i.e. via the old
namespace) if it's actually running, and start the new one (only if we
found and killed the old one).

Upgrade to ezbake 2.1.0 so that it'll include
fa844980a60e105270e0383ef1ddc4e8e3173d15, allowing it to accept our
shell code.

These changes can be removed in (at least) version 8 since we don't allow
skipping major versions, and so by that time, you will have upgraded
to at least one version with the fix.  (Of course you'll likely have
rebooted long before then, which will also "fix" the problem.)

While we're there, add a bit of quoting to some other
arguments/expansions.

[1] ezbake currently embeds the main namespace in various packaging
    scripts, and uses pgrep to find the server to kill, searching in
    part for that namespace, which originates in project.clj's
    [:lein-ezbake :main-namespace]).