(maint) Remove unneeded catalog inputs relationship
The relationships map is used to provide implicit subqueries to other
query entities. The default is to allow implicit subqueries by certname,
which is provided by the `certname-relations` map. The only reason to
add anything to the map is if you'd like it to support an implicit
subquery to an entity that doesn't return certname, like the environment
entity, or has duplicate entries for a certname, such as an implicit
subquery between fact_paths and fact_contents.