PDB-4487 - Constructing a new SSL context if we don't have local certs (aka masterless)
As part of work to migrate to Puppet::SSL::SSLContext, we have unintentionally broken
comapability with standalone/masterless nodes, who don't have a local CA or cert chain
to validate.

This commit introduces a new configuration variable (name TBC), and uses it to either
inherit from Puppet's SSL context (if we are running with a master and certificates)
or allows simple CA/CRL validation in the event that the PuppetDB server is using
either self-signed or commodity certificates.

Handoff to Puppet for these settings is via the configuration options `localcacert`
and `certificate_revocation` - if using a commodity certificate, `localcacert` should
be pointed to your OS local trust-store, and `certificate_revocation` set to `false`. If
using a self-signed certificate, this needs to be provisioned to each standalone node.