(PDB-2634) Hash projection for (some) json fields
Add the ability to project json subtrees as top-level results for simple
queryable json fields (`inventory`'s `facts` and `trusted` fields, and
`resources` `parameters` field). This corresponds to the fields that
support dotted notation in query syntax.

Extraction is done with dot notation representing the json fact
path, an AST query for the OS family would look like
  ["from", "inventory",
    ["extract", "facts.os.family"]]

In order to project json subtrees the projection entry needs two more
keys, :projectable-json? true, and :field-type (:raw|:keyword). The
prior lets the query engine to accept extractions likes "facts.os", but
not things like "certname.foo", while the latter tells it how to unpack
the :field and properly assembly the subtree extraction.

It is the responsibility of the user to ensure that their json key
exists because PuppetDB will return json null for both a key that exists
and is null and a key that does not exist (such as for a misspelling).