(maint) Change HikariCP FailFast to FailTimeout
The failfast property of HikariCP is deprecated and replaced with a
failtimeout.

Our previous setting of failfast=false would not error when the
minimum number of connections could not be obtained. The new fail
timeout setting of less than zero will mean the pool will start
immediately, and it will not timeout if it can't establish the minimum
number connections.