(PDB-4511) Don't require spaces in dotted projections
A whitespace character is currently required to terminate a dotted field
in PQL (see PDB-3284). Because it has been a requirement for multiple
major versions we are waiting for a major version to "break" it even
though we consider it a bug fix.

It would break a query like the one below, which uses a nonsensically
named "custom_param=", but we don't want to break existing queries in
a non-major version bump.

    resource { parameters.custom_param= = "Debian"}

Since this only affects the characters , and ] which are found in PQL
projection syntax, but not in the PQL query operators (= ~ etc), this
commit prevents a bug from propogating into a new feature and is
not a breaking change.