(PDB-3284) Make PQL !~<=>,] not need leading ws
Previously, due to greedy Java regexes, the special character operators
such as =, >, >=, ~, etc and field terminators like , and ] all
required a leading space. This makes the leading space optional.
The goal is that only spaces required are around the word operators
such as 'in', 'and', and 'or'.

This will break a query like the one below, which uses a nonsensically
named "custom_param=".

    resource { parameters.custom_param= = "Debian"}

We believe this to be a low risk thing to break, and unintentional
functionality as a result of the bug and therefore not a breaking
change.