(PDB-4502) Add a custom header for uncompressed command size
The 'Content-Length' header isn't sent when commands are compressed in the
terminus. As a result the PDB size metrics aren't updated and the
'max-command-size' configuration option which relies on 'Content-Length'
doesn't function as intended.

This commit adds a custom 'X-Uncompressed-Length' header to requests
sent from the terminus. The value of this header is the size in bytes of
the command before it is compressed. This value is used to update
the size metric and to decide when to discard commands based on
'max-command-size'.