(PDB-4509) Throw error for invalid database configuration
The postgres configuration 'standard_conforming_strings' is required in
(at least) two of our migrations, and is likely also required in some of
the query code. When 'on', it interprets a single backslash as a literal
character inside single quoted strings.

If standard_conforming_strings is 'off' a new install of PuppetDB will
fail to start up with an error 'invalid byte sequence for encoding
"UTF8": 0x00', this will instead print a more explanatory error to the
user.