(PDB-4513) Set correct cipher-suites setting
When we migrated to clj-parent 4.0.0 and then again in 4.2.0, the
default cipher-suites set by trapperkeeper were restricted. This could
occasionally result in puppetserver failing to connect with puppetdb.

This sets our default cipher-suite to match that of puppetserver, which
can be found
[here](https://github.com/puppetlabs/puppetserver/blob/master/ezbake/config/conf.d/webserver.conf).

We omit the cipher SSL_CK_DES_192_EDE3_CBC_WITH_SHA that puppetserver
supplies. If you want to continue to use SSLv3 (please don't), you need
to add that cipher suite to the list.