(PDB-4547) Add wireformat validation for deactivate node cmds
Prior to this commit it was possible to submit an invalid deactivate node
command that would fail during processing get retried multiple times but
remain in the stockpile queue and never make it to the dlo. This commit
adds validation for this command type so invalid commands will be rejected
before processing is attempted and sent to the dlo.