(PDB-4506) Don't hang after C-c (SIGINT) during migrations
Don't call System/exit when a migration fails.  Instead, just throw an
exception that will eventually be caught by trapperkeeper and initiate
the shutdown.  The System/exit was the cause of the hangs.

To reproduce the original problem, you can just add a Thread/sleep to
one of the migrations and hit C-c while pdb is sleeping.

Move the SQLException unraveling to the top level of migrate! so that
it'll cover the entirety of the migrations, and adjust it to unravel
the whole chain rather than just the first link.