(PDB-4550) Restore post-migration table analysis
Migrations can return an optional list of tables that should be vacuum
analyzed after all the migrations.  It has to come after since
vacuuming has to be done outside of any transaction.

I believe this support was originally added because we found that the
planner was constructing very poor plans for some queries because the
value_types table, which was only 5 rows and "never" changes, was
never analyzed.  Manually analyzing it substantially improved the
query plans.  The reason it was never analyzed is because by default
postgresql only analyzes tables above a certain size, or after a
certain amount of change.

However, it looks like this feature was broken/dropped during the
jsonb fact storage conversion, so restore it, but adjust it to make
sure not to try to examine intermediate tables that have been obviated
by the end of the migrations.