(maint) Explicitly delete reports/resource_events when deleting a certname
Prior to this commit we relied on the ON DELETE CASCADE triggers to clean up
reports and resource_events when deleting all data for a certname. This commit
makes the deletion of the report and resource_event rows for a certname explicit
leaving less work for the trigger to do when the certname's row is deleted from
the certnames table.