(PDB-4552) Add support for checking x-pdb-sync-ver header
This commit adds checking for the x-pdb-sync-ver header which is sent from
all pe-puppetdb sync requests. If there is a mismatch detected the incoming
sync request will receive a 409 conflict response until it is upgraded and
starts emitting a matching sync version. This is needed during pe-puppetdb
ha-sync to detect and avoid syncing with a pdb that has been recently upgraded
in a way that's not backwards compatible.