(PDB-4555) Retry settings validation db connection
This was found in the testing of the PuppetDB container, which starts
PuppetDB and PostgreSQL closer together than you'd traditionally see in
a VM or baremetal install.

The bug was introduced by PDB-4509 in 8152c4dd9, which performs the
database query 'show all;' outside the retry loop for
SQLTransientException. If that query encounters an error in its
connection, instead of retrying the connection it would error and
shutdown PuppetDB with the following message
    ERROR [p.t.internal] Error during service start!!!
    java.sql.SQLTransientConnectionException: PDBMigrationsPool -
    Connection is not available, request timed out after 3011ms.

An SQLTransientException indicates
    that the connection operation that failed might be
    able to succeed if the operation is retried without any
    application-level changes.

By moving into the retry loop for connection exceptions the error will
now be caught and retried with the same error message as migrations,
    ERROR [p.p.c.services] Will retry database connection after
    temporary failure:
    java.sql.SQLTransientConnectionException: PDBMigrationsPool -
    Connection is not available, request timed out after 3008ms.