(PDB-4512) Always list query parameters for routes
Intead of adding the paging parameters implicitly in
handlers/extract-query (not to be confused with http/extract-query),
depending on the arity, remove that arity from and just have every
route explicitly specify its param specs.

Add a merge-param-specs so we can easily combine param sets, and
create a "typical-params" set that includes all the params that were
being implicitly added by the two extract-queries (excepting
include_package_inventory -- we'll handle that later).

Aside from hopefully being a bit easier to follow, this prepares for
the addition of truly global options.  It's perhaps worth noting that
at the moment, not all routes are wrapped by extract-query.