(PDB-3417) Normalize all command metric names
Prior to this commit duplicate metrics could be created for incoming commands.
When the create-metrics-for-command! func was called with the query param
command name directly it would create a metric entry for "store_report" when all
other metrics updates used the normalized name "store report". This commit fixes
this issue for all command types so the normalized name is now used everywhere.