(PDB-4590) store reports with no events properly
A report with no events does not get the latest_report pointer in its
certname entry set. This occurred because of a nesting error.

The schema for reports does not allow an empty set of events, but that
was only used at testing time. In production, this happens frequently
(due to truncation of the resource events table, and other conditions.)