(PDB-4606) partition the reports table
Partition the reports table on producer_timestamp. This creates
a partition per day, and allows garbage collection to occur by
dropping expired tables. This follows the same pattern as PDB-4420.

The report hash has been changed to remove resource_events from the
hash calculation. The resource_events table can be garbage collected
separately from the reports table, and if the hash is still including
the resource_events, we end up with duplicate reports during a pdb
sync. The hash is rewritten during the migration.

There is no trigger to direct inserts to the appropriate partition.
This is because the usage of the reports table relies on the behavior
of a RETURNING clause on the INSERT INTO statement. You cannot rewrite
the query and combine it with a RETURNING clause. All INSERTs occur
directly into the appropriate partition.

During garbage collection, an UPDATE statement is ran across the
certnames table to ensure latest_report_id is set to null when the
report is deleted. Previously, this relied on a cascade during the
delete.